+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Upcoming version
+
 **Added:**
 - Network Service - Added Network Security Rule.
+- Network Service - Added Get Available IP Addresses count in Subnet
 
 ## 0.0.6